facedetection/requirements.txt
torchvision
pytorch
torchsummary
tqdm
opencv-python
pillow
matplotlib
scikit-image
